backend/pyproject.toml
[tool.poetry]
name = "Right Tools API"
version = "0.1.0"
description = "Right Tools API"
authors = ["Roman Sotnikov"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
drf-yasg = "^1.20.0"
djangorestframework-simplejwt = "^5.1.0"
psycopg2-binary = "^2.9.3"
Pillow = "^9.1.0"
gunicorn = "^20.1.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
